Move Product page component
