+++ src/components/Footer/Footer.module.css
@@ -4,5 +4,8 @@ footer {
 	color: var(--primary);
 	text-align: left;
 	display: flex;
-	align-items: center;
+	width: 100%;
+	padding-top: 13px;
+	padding-bottom: 13px;
+	padding-left: 40px;
 }
