+++ src/main.jsx
@@ -7,7 +7,7 @@ import {
 import App from './App.jsx'
 import ErrorPage from "./error-page";
 import './index.css'
-import Root from './pages/Root.jsx';
+import Root from './pages/RootLayout/Root.jsx';
 import ProductPage from './pages/ProductPage.jsx';
 
 // Root is used to render the basic layout, header and footer for all children elements.
