+++ src/pages/RootLayout/Root.module.css
@@ -1,0 +1,12 @@
+.rootLayout {
+	display: flex;
+	flex-direction: column;
+	min-height: 100%;
+	width: 100%;
+	align-items: center;
+	justify-content: center;
+}
+
+.layoutBody {
+	flex-grow: 1;
+}
